feat: tableGroup에서 order 의존성 제거
