+++ src/main/java/kitchenpos/table/application/OrderTableValidator.java
@@ -3,16 +3,22 @@ package kitchenpos.table.application;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.application.OrderValidator;
 import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 public class OrderTableValidator implements OrderValidator {
+    private final OrderRepository orderRepository;
     private final MenuRepository menuRepository;
     private final OrderTableRepository orderTableRepository;
 
-    public OrderTableValidator(MenuRepository menuRepository, OrderTableRepository orderTableRepository) {
+    public OrderTableValidator(OrderRepository orderRepository, MenuRepository menuRepository, OrderTableRepository orderTableRepository) {
+        this.orderRepository = orderRepository;
         this.menuRepository = menuRepository;
         this.orderTableRepository = orderTableRepository;
     }
@@ -28,6 +34,14 @@ public class OrderTableValidator implements OrderValidator {
                 .orElseThrow(() -> new IllegalArgumentException("주문 테이블을 찾을 수 없습니다."));
         if (orderTable.isEmpty()) {
             throw new IllegalArgumentException("빈 테이블은 주문을 할 수 없습니다.");
+        }
+    }
+
+    @Override
+    public void validateUngroup(List<Long> orderTableIds) {
+        if (orderRepository.existsByOrderTableIdInAndOrderStatusIn(
+                orderTableIds, OrderStatus.findNotCompletionStatus())) {
+            throw new IllegalArgumentException("조리 또는 식사 중인 테이블은 단체 지정을 해제할 수 없습니다.");
         }
     }
 }
