+++ src/test/java/kitchenpos/table/application/OrderTableValidatorTest.java
@@ -3,6 +3,8 @@ package kitchenpos.table.application;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import org.junit.jupiter.api.Test;
@@ -16,6 +18,7 @@ import java.util.List;
 import java.util.Optional;
 
 import static kitchenpos.application.OrderServiceTest.orderedTime;
+import static kitchenpos.application.TableGroupServiceTest.createOrderTableIds;
 import static kitchenpos.domain.OrderLineItemsTest.createDuplicateOrderLineItems;
 import static kitchenpos.domain.OrderLineItemsTest.createOrderLineItems;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -27,6 +30,8 @@ class OrderTableValidatorTest {
     private MenuRepository menuRepository;
     @Mock
     private OrderTableRepository orderTableRepository;
+    @Mock
+    private OrderRepository orderRepository;
     @InjectMocks
     private OrderTableValidator orderTableValidator;
 
@@ -65,5 +70,18 @@ class OrderTableValidatorTest {
                 orderTableValidator.validateOrder(order)
         ).isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("빈 테이블은 주문을 할 수 없습니다.");
+    }
+
+    @Test
+    void 조리_또는_식사_중인_주문_테이블이_있을_경우_단체_지정을_해제할_수_없다() {
+        // given
+        given(orderRepository.existsByOrderTableIdInAndOrderStatusIn(createOrderTableIds(), OrderStatus.findNotCompletionStatus()))
+                .willReturn(true);
+
+        // when & then
+        assertThatThrownBy(() ->
+                orderTableValidator.validateUngroup(createOrderTableIds())
+        ).isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("조리 또는 식사 중인 테이블은 단체 지정을 해제할 수 없습니다.");
     }
 }
